Fix repo shape after retention merge
